Add subscriptionInfo component
